backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
networkx==3.2.1
httpx==0.24.1
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
python-multipart==0.0.6
aiofiles==23.2.1
